Moved i18n from modules/contrib to modules/patched, applied a patch, and added a minor fix to the advanced site management view while at it.  Partially resolves issue #1787.
This missing translation is a known bug in Drupal.  I've applied a patch from http://drupal.org/node/346028#comment-1765998 to translate the list of icons beside the site names.  However, there is currently no working patch available online to translate the names in the selection box at the top.  Writing one will take a while, but this partial fix will improve the situation, so we can apply it in the meantime.